finish

+++ app/Http/Controllers/ListingController.php
@@ -43,6 +43,8 @@ class ListingController extends Controller
             $formFields['logo'] = $request->file('logo')->store('logos', 'public');
         }
 
+        $formFields['user_id'] = auth()->id();
+
         Listing::create($formFields);
 
         return redirect('/')->with('message', 'Listing created successfully!');
@@ -59,6 +61,12 @@ class ListingController extends Controller
     // Update Listing Data
     public function update(Request $request, Listing $listing) 
     {
+
+        // Make sure logged in user is the owner of the listing
+        if($listing->user_id != auth()->id()) {
+            abort(403, 'Unoauthorized action');
+        }
+
         $formFields = $request->validate([
             'title' =>'required',
             'company' => ['required'],
@@ -81,8 +89,19 @@ class ListingController extends Controller
 
     // Delete Listing
     public function destroy(Listing $listing) {
+        // Make sure logged in user is the owner of the listing
+        if($listing->user_id != auth()->id()) {
+            abort(403, 'Unoauthorized action');
+        }
+        
         $listing->delete();
         return redirect('/')->with('message', 'Listing deleted successfully!');
     }
 
+    // Manage Listings
+    public function manage() {
+        return view('listings.manage', [
+            'listings' => auth()->user()->listings()->get()
+        ]);
+    }
 }
